refactor(dup): Isolate simple flow into a dedicated module
