+++ apps/astarte_data_updater_plant/test/astarte_data_updater_plant/dup_producer_properties_test.exs
@@ -17,62 +17,32 @@
 #
 
 defmodule Astarte.DataUpdaterPlant.ProducerPropertiesTest do
-  use ExUnit.Case, async: true
+  use Astarte.Cases.Data, async: true
+  use Astarte.Cases.AMQP
+
   import Mox
+  import Astarte.Helpers.DataUpdater
 
   alias Astarte.DataUpdaterPlant.DatabaseTestHelper
-  alias Astarte.DataUpdaterPlant.AMQPTestHelper
   alias Astarte.Core.Device
   alias Astarte.DataUpdaterPlant.DataUpdater
   alias Astarte.DataAccess.Realms.Realm
   alias Astarte.DataAccess.Repo
   alias Astarte.DataAccess.Realms.IndividualProperty
   alias Astarte.DataAccess.Realms.Interface
-  alias Astarte.Core.Triggers.SimpleEvents.PathRemovedEvent
-  alias Astarte.Core.Triggers.SimpleEvents.SimpleEvent
   alias Astarte.Core.Device
   alias Astarte.DataAccess.Realms.IndividualDatastream
 
   alias Astarte.Core.CQLUtils
 
   import Ecto.Query
+  require Logger
 
   setup :verify_on_exit!
 
-  setup do
-    realm_string = "autotestrealm#{System.unique_integer([:positive])}"
-    {:ok, _keyspace_name} = DatabaseTestHelper.create_test_keyspace(realm_string)
-
-    on_exit(fn ->
-      DatabaseTestHelper.destroy_local_test_keyspace(realm_string)
-    end)
-
-    helper_name = String.to_atom("helper_#{realm_string}")
-
-    consumer_name = String.to_atom("consumer_#{realm_string}")
-
-    {:ok, _pid} = AMQPTestHelper.start_link(name: helper_name, realm: realm_string)
-
-    {:ok, _consumer_pid} =
-      AMQPTestHelper.start_events_consumer(
-        name: consumer_name,
-        realm: realm_string,
-        helper_name: helper_name
-      )
-
-    {:ok, %{realm: realm_string, helper_name: helper_name}}
-  end
-
-  test "producer properties are correctly set", %{
-    realm: realm,
-    helper_name: helper_name
-  } do
-    AMQPTestHelper.clean_queue(helper_name)
-
-    keyspace_name = Realm.keyspace_name(realm)
+  setup_all %{realm_name: realm_name} do
     encoded_device_id = "f0VMRgIBAQAAAAAAAAAAAA"
     {:ok, device_id} = Device.decode_device_id(encoded_device_id)
-
     received_msgs = 45000
     received_bytes = 4_500_000
     existing_introspection_map = %{"com.test.LCDMonitor" => 1, "com.test.SimpleStreamTest" => 1}
@@ -84,14 +54,37 @@ defmodule Astarte.DataUpdaterPlant.ProducerPropertiesTest do
       groups: ["group1"]
     ]
 
-    DatabaseTestHelper.insert_device(realm, device_id, insert_opts)
+    DatabaseTestHelper.insert_device(realm_name, device_id, insert_opts)
+    test_process = self()
+
+    Astarte.DataUpdaterPlant.RPC.VMQPlugin.ClientMock
+    |> Mox.stub(:delete, fn %{realm_name: ^realm_name, device_id: ^encoded_device_id} ->
+      send(test_process, :producer_properties_message_received)
+      :ok
+    end)
+
+    setup_data_updater(realm_name, encoded_device_id)
+
+    %{
+      device_id: device_id,
+      encoded_device_id: encoded_device_id,
+      received_msgs: received_msgs,
+      received_bytes: received_bytes
+    }
+  end
+
+  test "Unset values from interface properties", %{
+    realm: realm,
+    device_id: device_id,
+    encoded_device_id: encoded_device_id
+  } do
+    keyspace_name = Realm.keyspace_name(realm)
 
     data =
       <<0, 0, 0, 98>> <>
         :zlib.compress("com.test.LCDMonitor/time/to;com.test.LCDMonitor/weekSchedule/10/start")
 
     timestamp_us_x_10 = make_timestamp("2017-10-09T14:00:32+00:00")
-    timestamp_ms = div(timestamp_us_x_10, 10_000)
 
     DataUpdater.handle_control(
       realm,
@@ -125,28 +118,12 @@ defmodule Astarte.DataUpdaterPlant.ProducerPropertiesTest do
     )
 
     DataUpdater.dump_state(realm, encoded_device_id)
-    {remove_event, remove_headers, _meta} = AMQPTestHelper.wait_and_get_message(helper_name)
 
-    assert remove_headers["x_astarte_event_type"] == "path_removed_event"
-    assert remove_headers["x_astarte_device_id"] == encoded_device_id
-    assert remove_headers["x_astarte_realm"] == realm
-
-    assert :uuid.string_to_uuid(remove_headers["x_astarte_parent_trigger_id"]) ==
-             DatabaseTestHelper.fake_parent_trigger_id()
-
-    assert :uuid.string_to_uuid(remove_headers["x_astarte_simple_trigger_id"]) ==
-             DatabaseTestHelper.path_removed_trigger_id()
-
-    assert SimpleEvent.decode(remove_event) == %SimpleEvent{
-             device_id: encoded_device_id,
-             event:
-               {:path_removed_event,
-                %PathRemovedEvent{interface: "com.test.LCDMonitor", path: "/time/from"}},
-             timestamp: timestamp_ms,
-             parent_trigger_id: DatabaseTestHelper.fake_parent_trigger_id(),
-             realm: realm,
-             simple_trigger_id: DatabaseTestHelper.path_removed_trigger_id()
-           }
+    # {remove_event, remove_headers, _meta} = AMQPTestHelper.wait_and_get_message(amqp_consumer)
+    #
+    # assert remove_headers["x_astarte_event_type"] == "path_removed_event"
+    # assert remove_headers["x_astarte_device_id"] == encoded_device_id
+    # assert remove_headers["x_astarte_realm"] == realm
 
     endpoint_id = retrieve_endpoint_id(realm, "com.test.LCDMonitor", 1, "/time/from")
 
@@ -239,17 +216,5 @@ defmodule Astarte.DataUpdaterPlant.ProducerPropertiesTest do
     {:ok, endpoint_id} = Astarte.Core.Mapping.EndpointsAutomaton.resolve_path(path, automaton)
 
     endpoint_id
-  end
-
-  defp gen_tracking_id() do
-    message_id = :erlang.unique_integer([:monotonic]) |> Integer.to_string()
-    delivery_tag = {:injected_msg, make_ref()}
-    {message_id, delivery_tag}
-  end
-
-  defp make_timestamp(timestamp_string) do
-    {:ok, date_time, _} = DateTime.from_iso8601(timestamp_string)
-
-    DateTime.to_unix(date_time, :millisecond) * 10000
   end
 end
