+++ apps/astarte_data_updater_plant/test/astarte_data_updater_plant/dup_unset_properties_test.exs
@@ -17,8 +17,11 @@
 #
 
 defmodule Astate.DataUpdaterPlant.UnsetTest do
-  use ExUnit.Case, async: true
+  use Astarte.Cases.Data, async: true
+  use Astarte.Cases.AMQP
+
   import Mox
+  import Astarte.Helpers.DataUpdater
 
   alias Astarte.DataUpdaterPlant.DatabaseTestHelper
   alias Astarte.DataUpdaterPlant.AMQPTestHelper
@@ -36,41 +39,9 @@ defmodule Astate.DataUpdaterPlant.UnsetTest do
 
   setup :verify_on_exit!
 
-  setup do
-    realm_string = "autotestrealm#{System.unique_integer([:positive])}"
-    {:ok, _keyspace_name} = DatabaseTestHelper.create_test_keyspace(realm_string)
-
-    on_exit(fn ->
-      DatabaseTestHelper.destroy_local_test_keyspace(realm_string)
-    end)
-
-    helper_name = String.to_atom("helper_#{realm_string}")
-
-    consumer_name = String.to_atom("consumer_#{realm_string}")
-
-    {:ok, _pid} = AMQPTestHelper.start_link(name: helper_name, realm: realm_string)
-
-    {:ok, _consumer_pid} =
-      AMQPTestHelper.start_events_consumer(
-        name: consumer_name,
-        realm: realm_string,
-        helper_name: helper_name
-      )
-
-    {:ok, %{realm: realm_string, helper_name: helper_name}}
-  end
-
-  test "Unset values from interface properties", %{
-    realm: realm,
-    helper_name: helper_name
-  } do
-    AMQPTestHelper.clean_queue(helper_name)
-
-    encoded_device_id = "f0VMRgIBAQAAAAAAAAAAAA"
-    keyspace_name = Realm.keyspace_name(realm)
-    {:ok, device_id} = Device.decode_device_id(encoded_device_id)
-    volatile_changed_trigger_id = :crypto.strong_rand_bytes(16)
-
+  setup_all %{realm_name: realm_name} do
+    device_id = Device.random_device_id()
+    encoded_device_id = Device.encode_device_id(device_id)
     received_msgs = 45000
     received_bytes = 4_500_000
     existing_introspection_map = %{"com.test.LCDMonitor" => 1, "com.test.SimpleStreamTest" => 1}
@@ -82,7 +53,35 @@ defmodule Astate.DataUpdaterPlant.UnsetTest do
       groups: ["group1"]
     ]
 
-    DatabaseTestHelper.insert_device(realm, device_id, insert_opts)
+    DatabaseTestHelper.insert_device(realm_name, device_id, insert_opts)
+    test_process = self()
+
+    Astarte.DataUpdaterPlant.RPC.VMQPlugin.ClientMock
+    |> Mox.stub(:delete, fn %{realm_name: ^realm_name, device_id: ^encoded_device_id} ->
+      send(test_process, :unset_properties_message_received)
+      :ok
+    end)
+
+    setup_data_updater(realm_name, encoded_device_id)
+
+    %{
+      device_id: device_id,
+      encoded_device_id: encoded_device_id,
+      received_msgs: received_msgs,
+      received_bytes: received_bytes
+    }
+  end
+
+  test "Unset values from interface properties", %{
+    realm: realm,
+    amqp_consumer: amqp_consumer,
+    device_id: device_id,
+    encoded_device_id: encoded_device_id
+  } do
+    AMQPTestHelper.clean_queue(amqp_consumer)
+
+    keyspace_name = Realm.keyspace_name(realm)
+    volatile_changed_trigger_id = :crypto.strong_rand_bytes(16)
 
     assert DataUpdater.handle_delete_volatile_trigger(
              realm,
@@ -90,15 +89,15 @@ defmodule Astate.DataUpdaterPlant.UnsetTest do
              volatile_changed_trigger_id
            ) == :ok
 
-    DataUpdater.handle_data(
-      realm,
-      encoded_device_id,
-      "com.test.LCDMonitor",
-      "/weekSchedule/10/start",
-      <<>>,
-      gen_tracking_id(),
-      make_timestamp("2017-10-09T15:10:32+00:00")
-    )
+    # DataUpdater.handle_data(
+    #   realm,
+    #   encoded_device_id,
+    #   "com.test.LCDMonitor",
+    #   "/weekSchedule/10/start",
+    #   <<>>,
+    #   gen_tracking_id(),
+    #   make_timestamp("2017-10-09T15:10:32+00:00")
+    # )
 
     DataUpdater.dump_state(realm, encoded_device_id)
 
@@ -141,17 +140,5 @@ defmodule Astate.DataUpdaterPlant.UnsetTest do
     {:ok, endpoint_id} = Astarte.Core.Mapping.EndpointsAutomaton.resolve_path(path, automaton)
 
     endpoint_id
-  end
-
-  defp make_timestamp(timestamp_string) do
-    {:ok, date_time, _} = DateTime.from_iso8601(timestamp_string)
-
-    DateTime.to_unix(date_time, :millisecond) * 10000
-  end
-
-  defp gen_tracking_id() do
-    message_id = :erlang.unique_integer([:monotonic]) |> Integer.to_string()
-    delivery_tag = {:injected_msg, make_ref()}
-    {message_id, delivery_tag}
   end
 end
